LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/moving-between-documents/after-prepare-iframe-parse-error-external-module-expected.txt

CONSOLE MESSAGE: SyntaxError: Left side of assignment is not a reference.


PASS Sanity check around top-level Window
PASS Eval: Move parse-error external module script to iframe after-prepare
PASS <script> load: Move parse-error external module script to iframe after-prepare
FAIL <script> error: Move parse-error external module script to iframe after-prepare assert_unreached: Script error evennt fired unexpectedly Reached unreachable code
PASS window error: Move parse-error external module script to iframe after-prepare
